requirements.txt
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
MarkupSafe==2.1.2
python-digitalocean==1.17.0
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.12.0
